Update patch set 1
Patch Set 1:

@YueChen
I agree that the luma tx size is 8x8 in supertx mode, however the failing test occurs when using 4:2:2 chroma subsampling. The chroma tx size is 4x4 but as far as we understand there are 2 blocks that need to be tokenized.  Using the EndToEnd test the original code works fine in 4:2:0 or 4:4:4, but fails in 4:2:2 and 4:4:0.

Patch-set: 1
Label: Code-Review=0