Update patch set 1
Patch Set 1:

I think a reasonable alternative would be to hard-code 16384 as the size restriction and make it the default. Then we don't have to pass tricky numbers down through configure. --enable-size-limit (default) would set it to 16384x16384, and --disable-size-limit would either disable the check or set it to 64kx64k

Patch-set: 1