Allowing MV overflow
If the bitstream codes a delta that when added to the predictor results
in a value over MV_IN_USE_BITS bits, the value is wrapped as if it is an
overflow. That way no vector can be coded that is larger than expected.

This prevents any bitstream from being illegal this way, thereby
reducing the need for runtime conformance checking. It also gives an
encoder a cheap way to code a delta that can reverse the direction of a
large vector, should it ever encounter content that would benefit from
that.

Change-Id: I16ab13f65a2d32ce9749596dcd144ef7dc3f2ebe