Update patch set 4
Patch Set 4:

The impact was small in my tests.

For sub hd image formats the minimum is now 4x4 anyway. I may look to change this so that it is content type specific if I can see a speed reason to do so on non animated content.

For larger formats only a small % of the blocks are at the active edge and for those there is then a small % increase in cost.

Patch-set: 4
Label: Code-Review=0