Update patch set 4
Patch Set 4:

> @jzern
 > 
 > Follow up the prefix topic. Could you point an example use case? We
 > can discuss the naming conventions later. Thanks.

Use case or an example of adding a prefix?
The benefit to being consistent with inlines and the C source is it's clear to the reader from which module they come from and avoids a patchwork of changes like this. To me only the simplest tests would be allowed to be unprefixed, e.g., is_...().

Patch-set: 4