Fix to non-rd variance partition selection
Only test for using golden as reference for variance partition
selection if it used as a reference for that frame.

For temporal layers, golden may not be a reference on a given frame,
even though it was for some previous frame. If it is not a current reference,
no need to check/use it for partition selection.

Change-Id: I6b0f2bd36aebbb5903077c9a0a66d80f1de9a7b1