Fix to golden update in cyclic_refresh.
For one pass CBR: only check for updating refresh_golden
if ext_refresh_frame_flags_pending is not set (i.e., == 0).

This is to prevent changing refresh_golden if the user
supplies the reference flags. This change moves the resetting
of ext_refresh_frame_flags_pending = 0 to after the encode_loop
call to achieve this.

Change-Id: I4d87b3e705ba43f243667e367503b585c61e2a54