Fix to dynamic resize mode under change_config().
If the encoder dynamic resize is triggered and
change config() is then called, it will reset the
current (resized) codec width/height back to the
the config (unresized) width/height (which will then
prevent the resizing action in encoder_loop).

Avoid this by checking for a change in the config width/height
before resetting the codec width/height.

Change-Id: Id9d50c0ee8a943abe4b6c72bbaa02d9696f93177