vp10: fix entropy counts for the hp bit.
The counts didn't take usehp into account, which means that if the scope
of the refmv is too large for the hp bit to be coded, the value (always
1) is still included in the stats. Therefore, the final counts will not
reflect the entropy of the coded bits, but rather the entropy of the
combination of coded bits and the implied value (which is always 1). Fix
that by only including counts if the hp bit is actually coded.

See issue 1060.

Change-Id: I19a3adda4a8662a05f08a9e58d7e56ff979be11e