No scale 32x32 experiment
An experiment to use an 8 times orthogonal scaled 32x32
transform, rather than the 4-times orthogonal transform
used in VP9.

The new 32x32 transform is used only when the residue is
within the range -128 to 128, thus ensuring that the
coefficients fir in 16 bits.

Results on derflr and hevcmr are slightly down but within
noise.

This is an experiment for reference purposes only, to see
if there is any major imapct on the scaling, and as of now
there does not seem to be any.

Change-Id: I4d4960b70c6092340162822a9bf4bca8698c210d