Update patch set 3
Patch Set 3:

I would assume a speed penalty for checking for zero. The assert is only active in debug builds.

The first case with num_values might be able to check 0 inside the function, I'd assume it would be a wash in that case.

The second looks like the minimum value is 8 so it's safe?

Patch-set: 3