Set sizes of function symbols on ELF targets.
This commit changes no binary code, but adds macros to set the size of
function symbols in assembly files when building for targets the use the
ELF object format. The symbol size can be used by profilers (e.g.: perf,
callgrind) to associate instructions with function names when debug
information is not available. This change aims to enable generating
more informative profiles.

Writers of x86 assembler functions should issue the 'END' macro as the
last statement in a source file. Omitting this will not cause an error,
but might make the reports from profilers somewhat less informative.

Change-Id: Ib6c632a3ab23f2476691bdef3fdfdecf18be4aed