Only emit '%%branch_instr:' if necessary
There is a macro mechanism in x86inc.asm that emits a macro local label
after branch instructions to handle some micro-architecture
inefficiencies in older x86 processors. This pollutes the symbol table,
and yasm generates some debug info that is confusing for profilers. The
result is that instructions after these labels are associated to the
label (and not to the function they are contained in) in profiling
reports, which is unhelpful. With this patch, the problematic labels are
not generated if we know that the target processor does not suffer from
this inefficiency.

Change-Id: Ie6fed65beee4203ee6720618cf7d86d1b200ae49