Update patch set 8
Patch Set 8:

This has some quality effects, but it's hard for me (since I lack the ability to test this on larger data sets) to say whether that's positive or not and what it's caused by. I'm suspecting that the default tables between "former" intra/inter need to be merged based on some ratio between the two. Is there some way I can run tests myself?

Patch-set: 8