Update patch set 2
Patch Set 2:

The build failure was due to the version of yasm used by Jenkins (1.1.0 I presume) not recognizing 'pxor ymm5, ymm5' as a valid AVX instruction. yasm 1.2.0 accepts this. There is ambiguity in the Intel documentations on whether pxor ymm, ymm is in AVX or AVX2 only, so I reverted to using pxor xmm, xmm, relying on the vzeroupper in the beginning of the function to zero the top half of the ymm registers.

Patch-set: 2
Reviewer: Gerrit User 1137274 <1137274@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0