VP9_LPF_HORIZONTAL_8_DUAL_SSE2 optimization
The intrinsic version of that function caused gcc to generate a lot of fills and spills,
by writing it in plain assembly and making sure that instruction bigger then 8 bytes do not prevent from 2 instruction that are less then 8 bytes to retire in the same time.
The function level gain is 32% on 32bit and 4% on 64bit on Architecture based Atom.
credit for writing the function goes to Tamar Levy(tamar.levy@intel.com) and Ilya Albrekht (ilya.albrekht@intel.com)

Change-Id: If0a2a9fa552d5868ac9b0645d42709a0ebfc866d