Update patch set 3
Patch Set 3:

Thanks for the patch! Mozilla has signed the corporate CLA. We don't actually check the CLA programatically. Your first patch needs to be "shepherded" through. I won't be the one to approve this particular one but I've reuploaded it to get it through the tests and will add jzern@ to take a look. Apologies for any delay, but it is a (almost) holiday weekend.

Patch-set: 3