MB_LPF_HORIZONTAL_EDGE_W_SSE2_16 optimization
This function had a lot of stalls becuase of unecessary fills and spills to the stack,
and using instruction that are bigger then 8 bytes which on silvermont Architecture cause a stall.
The only way to optimize that function was to write it in plain assembly and not using intrinsic and
also using rbp instead of rbp when saving into the stack
(rsp use extra byte when the instrcution is decoded and may cause the instruction to be bigger then 8 byte).
The function level gain is 32% for both 32bit and 64bit. The user level gain is up to 2%.
credit for writing the function goes to Tamar Levy(tamar.levy@intel.com) and Ilya Albrekht (ilya.albrekht@intel.com)

Change-Id: I89d9b6536f056990554bc9efed9ef52ced20b3ab