Added a new experiment "prev_mvref"
This experiment is to trace the reference frame buffer update, both at
the encoder and at the decoder, so that we may identify whether the
reference frames for the current frame and those for the previously
coded frame are identical or not. If they are indeed identical, the
motion vectors of the co-located block in the previous frame may
provide a more accurate prediction for the motion vector of the
current block, hence being a better candidate as the motion vector
reference.

Change-Id: I79a117655b84f3a8dc9ef68b7a46019d6d0c1ee6