Alter loop filter search bias conditions.
Adjust the calculation of the bias used in the loop filter search
towards lower values. Remove adjustment that was applied
only in 2 pass based on the intra inter ratio (subject to review).

One side effect of the 2 pass inter intra ratio adjustment was that
1 pass encodes were using a much higher bias (2x-10x) than most 2
pass encodes which may have given rise to under filtering in 1 pass.

Change-Id: Ib63d7d48b253469b2918aea26a3758c06d52f3af