Do not update VAR_TX related context with SUPERTX
The encoder did not update left_txfm_context and above_txfm_context in
MACROBLOCKD (used for choosing the probability context for the vartx
split bits) when the supertx bit was set for a block. The deoder on the
other hand did update these for supertx blocks. The encoder used these
to compute the context counts, which the packer then uses to adapt it's
probabilities. This results in the packer and the decoder using
different probabilities.

This patch harmonizes the encoder and the decoder by making the packer
and decoder not update the mentioned context for supertx coded blocks.

Change-Id: I4aae96e625e3ef2dfc56365129fca97215af01df