Update patch set 3
Patch Set 3:

Nico Weber from chrome/clang helpfully pointed out a related presentation and bug [1][2]. The recommendation is to use /FIintrin.h for now. I hadn't thought of that, but it's simpler, so I'd prefer not to take this change since it can be worked around on the compiler side.

[1] https://docs.google.com/presentation/d/1oxNHaVjA9Gn_rTzX6HIpJHP7nXRua_0URXxxJ3oYRq0/edit?usp=sharing
[2] http://llvm.org/PR19898

Patch-set: 3