Update patch set 2
Patch Set 2:

In VP8 multithreading encoder/decoder, the process of a macroblock can only be started after its above macroblock is done, which guarantees the correct processing order. However, TSan couldn't understand that. I added synchronizations to thread-shared data to reduce the data race warnings, which causes unavoidable speed loss. But, several things were done to lower the overhead as much as possible.
1. Multiple mutexes were used, where 1 mutex was created for 1 macroblock row. So only the thread handling the above row (write) and the thread handling the current row (read) need to lock the mutex. Since the current macroblock has to wait for the above microblock - this won't cause much overhead.
2.x86_pause_hint();
  thread_sleep(0);
The above lines were used to implement the busy waiting mechanism. This was the original implementation in VP8, and was not modified in the fix.

Patch-set: 2
Label: Code-Review=0